requirements.txt
arrow==0.15.5
astroid==2.3.3
attrs==19.3.0
binaryornot==0.4.4
bleach==3.1.4
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
cookiecutter==1.7.0
docutils==0.16
future==0.18.2
idna==2.9
importlib-metadata==1.5.0
isort==4.3.21
Jinja2==2.11.1
jinja2-time==0.2.0
keyring==21.2.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.2.0
packaging==20.1
pkginfo==1.5.0.1
pluggy==0.13.1
pop==8.2
poyo==0.5.0
py==1.8.1
Pygments==2.6.1
pylint==2.4.4
pyparsing==2.4.6
pytest==5.3.5
python-dateutil==2.8.1
PyYAML==5.3
readme-renderer==25.0
requests==2.23.0
requests-toolbelt==0.9.1
six==1.14.0
tqdm==4.43.0
twine==3.1.1
typed-ast==1.4.1
urllib3==1.25.8
wcwidth==0.1.8
webencodings==0.5.1
wheel==0.34.2
whichcraft==0.6.1
wrapt==1.12.0
zipp==3.0.0
